# ⚡ Speedup merge