@magiconair 2022-12-08
Done. Merged to v1.8.7